changed resize specs
